Accept credentials for build and deploy as a JSON string
Right now only credentials cannot be hardcoded so we need to pass them
in as inputs to the build. Note that the values of the credentials are
base64 encoded. I don't like this and would rather have something more
susinctly just JSON escape them but couldn't find a simple way to do
this in bash.

- Build image now requires jq