Update global state when diff input is changed
Previously the input component owned the state of the text area. In
order to make this information more readily available to the rest of the
application we want it to change in line with what is in the text box.
Note that currently nothing outside of the input actually needs the
input however this is done for the general idea stated above as it seems
like a more ideal design.

- Split global state up into two individual state variables