Add temporary/unfinished test for diff input component
This test is weird and overly complicated. Particularly with the
rerenders that are necessary. I also am not a fan of the assertions
we have. Assertions on number of times a function has been called
aren't ideal. I'd prefer if we assert on actual values being what
is expected. Perhaps there is a way to use useState in this test