Refactor input component to not use useEffect
I'm not sure we were using useEffect() for the right reasons here.
We wanted something which ran when the post data was submitted and
had a change of that trigger the useEffect. However, we could just
avoid post data altogether by just calling fetch when the submit
button was pressed.

The main reason we were using useEffect() was so that when the
button was clicked when there were no changes to the post data then
it wouldn't submit. However, this is overloading the use of
useEffect() for reasons it isn't necessarily for.

Using it this way was also making this component much harder to
test.