Add test.sh script to run a full build and test of the project
The script can be run and it will essentially run the entire build
locally. The secrets will have to be filled out manually.

Since it runs the entire build it will also deploy to prod. This means
that certain lines in your local code should be commented out to prevent
this before running the script.