Serve website only via HTTPS
The website was being served up by HTTP only previously.

To serve it up via HTTPS we needed to put a CloudFront distribution in
front of the S3 bucket. We also needed to create a certificate.

Note that until conjob is also HTTPS calls to the backend won't work
since the browser will refuse to make a non-HTTPS request from an HTTPS
site.

Note that the site is still publically available if you directly go to
the buckets URL, but no user should do this as the address isn't
publically available.