Warm up backend endpoint before running cypress tests
Tests were timing out because the backend would be slow to respond;
likely because it didn't have the job it needed to run and was pulling
it down.

Instead of increasing the timeout for all our tests, we'll leave their
timeouts modest and instead fix the cause of a lot of the timeout issues
by making a request to the backend endpoint to warm it up so we know it
won't need to pull the image during a test.