Remove all files in S3 that aren't in the latest build
Each deploy we sync the files in our current build with what's in S3
currently. However, this means that there can be files left over in S3
from previous deploys. By deleting files that aren't in the current
build we ensure both contain the same exact files.

This change is spurred by CloudFlare seeming to cache old versions of
files even after they've been updated. The Last-Modified headers
returned by CloudFlare don't match up with the last modified time in S3,
so it's unclear why it wouldn't be sending new files. I must not have a
clear understanding of the cache policy.
It's possible that this change won't actually fix this problem since
changes in S3 don't even seem to have an effect on what CF returns, but
it's possible that it will fix it and this change should be made
regardless.