Disabled CloudFront caching
The latest version of changes were not showing up and it was difficult
to troubleshoot what the issue was. It seemed like it may have had
something to do with compressed versions of the files being cached.

In any case, this entirely disabled CF caching since we really don't
need it anyways.