Move CloudFront tf code from core module to test and prod envs
We were never associating the prod certs with the CF ditribution. This
forks the dist into two similar resources for prod and testing. The
difference being that testing uses the default cert, while prod uses one
created with ACM.

I think the only reason this was working before was that we had some
left over resources that were created manually, such as the cert, and
manually accociated them with the CF dist.