Wait for cert validation before creating CF dist
The CF dist has no natural dependency on the cert validation resource.
It only depends on the cert resource. So after it's created the CF dist
will start to be created, but will fail if the cert is not valid.

Now the CF dist resource will wait until the cert validation resource
has finshed before starting creation.