Don't specify domain name in alternatives
Specifying the root domain name in the alternatives list was causing the
cert to change on every subsequent application which also cause the
CloudFront distribution to try to recreate itself. The CF dist waits
forever to try to delete itself and I believe this is the root cause of
that.