Changes necessary for the @testing-library/user-event update
Updating to @testing-library/user-event required some changes to the
tests.

Every time we typed into the text area even though updateDiffInput is
called the actual value isn't updated since it's a mock. There was a
change to the testing lib which broke this test. I think it makes sense
that the way we were testing it was wrong though.