Update terraform backend-init for terraform upgrade
Changes made to the main tf configs needed to be pulled in here. These
changes are to specify versions explicitly and to pull out bucket
configuration into their own resource.

Note that we should maybe also import the version resource during the
build, but it seems to be fine to not. It will just "create" it each
time. We only need to do this for the bucket itself otherwise it will
fail to create it since it already exists, but this doesn't seem to be a
problem for configuration that's part of the bucket if that already
exists.