Remove tool and package version validations temporarily
These checks would fail the build if any of the tools or package
dependencies were not at their latest version.

However, it's difficult to keep up with new versions and I don't think
it should be failing the main build. I'd like to have a separate
validation for this that runs on a schedule separate from the main
build.