Support different build tasks
Adds support for specifying a build task when calling build.sh.

Consequences of this are that now we somewhat unfortunately have some
conditional logic in the build pipeline which I was trying to avoid, but
I think this is kinda unavoidable. It helps with calling test.sh because
previously I'd just manually comment out the deploy step which was
pretty hacky, though it doesn't address the git clone issue.

This change was initiated to support PRs and branches, mainly so we can
use things like renovate.