Use latest name for backend image
This was renamed a while ago and updated elsewhere in the code, but not
here. This URL is used to make sure the backend image is already loaded
once before the tests actually start. Since we were using the wrong URL
here it wasn't actually warming things up and this may have been causing
acceptance test failures due to timeouts.