Support using branches and alternative docker tags with build images
Our PRs were just using the latest build image. However, if there are
changes to the build system we would not be testing them.

We now build and tag a build image as part of PR builds and then use
that tag when running tests and such on the build image.