Only create renovate PRs if tests fail
Instead of opening PRs for every renovate change, it just creates the
branch, runs the tests, and then merges the changes if the tests
succeed. If the tests fail it still creates a PR.

This will hopefully reduce a lot of notification noise created from all
the renovate version bumps.